chore: rename clippy-token-color-table -> clippy-token-table-color
