+++ .changeset/khaki-emus-read.md
@@ -2,4 +2,4 @@
 '@nl-design-system-community/clippy-components': minor
 ---
 
-added clippy-token-color-table component
+added clippy-token-table-color component
