+++ packages/clippy-components/src/clippy-token-table-color/README.md
