+++ packages/clippy-components/src/clippy-token-table-color/fixtures.ts
