+++ packages/clippy-components/src/clippy-token-table-color/types.ts
